Database created. Generic Repository Design Pattern applied.

+++ Entities/Concrete/Car.cs
@@ -9,8 +9,8 @@ namespace Entities.Concrete
     {
         public int CarId { get; set; }
         public int BrandId { get; set; }
-        public short ColorId { get; set; }
-        public decimal ModelYear { get; set; }
+        public int ColorId { get; set; }
+        public int ModelYear { get; set; }
         public int DailyPrice { get; set; }
         public string Description { get; set; }
     }
